Add journalling template
This commit introduces a new `org-capture-templates` entry meant for
Journalling. It uses `file+olp+datetree` without heading argument and
thus automatically creates (or searches) the correct level 3 entry:

  * YEAR
  ** MONTH
  *** DAY

The entry itself is prefixed with four stars to enable multiple
headings on a single day, if I feel like it.

I doubt that I will use this template for a long while, but maybe try
it for some days. It has to start somewhere, right?

By the way, the `org-capture-template` is also customized on this
machine; the customizations are private and thus not included in the
repository. See also `(use-package emacs ...)`.