Use flex completion style and ignore case
I'm not yet 100% sold on the orderless package. It probably solves
this even better, but so far I'm trying to keep the number of packages
relatively low.

The `flex` completion style is closer to the original Ivy completion
style than emacs22. I still miss   space   driven   patterns, but on
the other hand they don't work on vanilla `completing-read` either. So
let's try to get used to partial-completion and flex instead!