Display a column indicator in the configuration
This commit enables display-fill-column-indicator-mode in the init.el,
which somewhat forces me to respect my self imposed limits on the
number of lines and characters.

It also takes yet another line as a victim. It's not necessary, sure,
but it's really helpful to immediately see lines that don't follow the
80 character limit.