early-init: move parts of configuration into early-init
This change cuts about 0.1s-0.2s of the original startup time However,
this is yet another file within `.emacs.d`, and therefore needs to
counted within the line limits. It has its own sublimit of 100 lines,
which might seem much, but then again , there are 10 SLOC, which is
only a fifth of the lines commited within the file.