emacs: show trailing whitespace only in prog and text derived modes
As a C-variable, `show-trailing-whitespace` is one of the... strange
ones. Even though it acts as a variable, setting it will change the
behavior of the current buffer, similar to a mode.

With "real" files, this is not an issue as usual, as trailing
whitespace is not desired there. But with buffers like `org-agenda`,
the trailing red line around elements that were not meant to show
their trailing whitespace via `show-trailing-whitespace` are not
really pleasing to the eye.

The hack via `read-only-mode-hook` kind of worked. But only kind
of. For some strange reasons, it sometimes misfired, for example in
M-x calendar. But all those special modes for user entry don't use
text-mode or prog-mode behind the scenes. For example,
`tabulated-list-mode` does not use `text-mode` nor `prog-mode`, but
`special-mode`.

Using the hooks in the appropriate modes that are meant for code/text
instead of the nuke option `setq-default` promises to be a lot less
error prone, so I'll keep it for the moment.