use-package: fix :bind to defer package loading
It turns out that (function <func>) inhibits use-package's deferred
load. Instead, use-package will immediately `(require)` the package
at the start of Emacs, even though it's not necessary.

Not all packages are affected, though. `org` mode for example still
only loaded on demand, which is fortunate, as this takes more than
3 seconds to load on a Windows machine. Still, changing the syntax
from `(key . #'func)` to `(key . func)` decreased the overall (hot)
`emacs-init-time` by 0.2s on Windows (from 1.3s to 1.1s).

Usually, I'd use `esup` for this kind of profiling, but it has been
defunct for a while on Windows. Tracing the functions on the other
hand has provided the required clues:

    (trace-background-function 'require "*trace:require*")
    (trace-background-function 'load "*trace:load*")

That showed me that some features were unexpectedly required.